server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.1",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.5",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.48",
    "typing>=3.10.0.0",
    "uvicorn>=0.41.0",
]
